+++ packages/examples-site/CHANGELOG.md
@@ -1,5 +1,34 @@
 # bigcommerce-examples-site
 
+## 0.7.0
+
+### Minor Changes
+
+- 8d399ef: Add BigDesign AI prompt nav item and updated home screen
+
+### Patch Changes
+
+- Updated dependencies [8d399ef]
+  - bigcommerce-design-patterns@0.11.1
+
+## 0.6.0
+
+### Minor Changes
+
+- 612d0ab: Add sidebar
+
+### Patch Changes
+
+- Updated dependencies [612d0ab]
+  - bigcommerce-design-patterns@0.11.0
+
+## 0.5.3
+
+### Patch Changes
+
+- Updated dependencies [c4b8704]
+  - bigcommerce-design-patterns@0.10.0
+
 ## 0.5.2
 
 ### Patch Changes